all function are included in feed page
